Restructure scripts folder
