+++ scripts/initialize.js
@@ -15,7 +15,7 @@ const {
 } = require("@solana/spl-token");
 const BN = require("bn.js");
 
-const { mintAuthority, user, mintKey } = require("./keys.js");
+const { mintAuthority, user, mintKey } = require("./setup/keys.js");
 const programId = new PublicKey("FdmChujE5rEhsvmTUvz1gbfoU3bKSWi52YuSqghNaDhj");
 const connection = new Connection("https://api.devnet.solana.com/");
