+++ scripts/setup/initTokens.js
