+++ scripts/setup/keys.js
